fix(all): rename module
